nfpm.native_libs.deb: split tests into separate files
